requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-import-export>=3.2.0
openpyxl>=3.1.0
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg[binary,pool]>=3.1.0
dj-database-url>=2.1.0
python-decouple>=3.8
